+++ Changelog.md
@@ -49,6 +49,7 @@ v2.37.0 (not published yet)
 - UI: Improve handling with Dolby Vision videos and manual cropping
 - UI: Improve Dolby Vision metadata handling
 - UI: Enhance "Frame Mismatch" error message
+- UI: Let "Frame Mismatch" ignore cutted "Copy/Mux" runs ([#1234](/../../issues/1234))
 - Update tools
     - ...
 - Update AviSynth+ plugins